backend/requirements.txt
# === Core Framework ===
fastapi==0.115.12
starlette==0.46.1
uvicorn==0.15.0
anyio==4.9.0
click==8.1.8
h11==0.14.0
sniffio==1.3.1
exceptiongroup==1.2.2
annotated-types==0.7.0

# === Firebase Integration ===
firebase-admin==6.2.0
CacheControl==0.14.2
cachetools==5.5.2
msgpack==1.1.0

# === Google / Gmail API & Auth ===
google-api-core==2.24.2
google-api-python-client>=2.81.0
google-auth>=2.16.2
google-auth-httplib2==0.2.0
google-auth-oauthlib>=1.0.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
rsa==4.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
uritemplate==4.1.1

# === Networking / HTTP ===
httpx==0.28.1
httpcore==1.0.7
requests>=2.28.2
requests-oauthlib==2.0.0
aiohttp>=3.8.4
aiodns>=3.0.0
idna==3.10
charset-normalizer==3.4.1
urllib3==2.3.0
certifi==2025.1.31
distro==1.9.0

# === Security / Encryption ===
cryptography==44.0.2
PyJWT==2.10.1
cffi==1.17.1
pycparser==2.22
oauthlib==3.2.2

# === Data Validation / Parsing ===
pydantic==1.10.7
pydantic_core==2.33.0
typing_extensions==4.13.0
typing-inspection==0.4.0
email-validator==1.1.3
python-multipart==0.0.5
python-dateutil==2.8.2
pyparsing==3.2.3

# === Utilities ===
python-dotenv==0.19.0
tqdm==4.67.1
numpy>=1.24.2
proto-plus==1.26.1
protobuf==5.29.4
openai==1.70.0
jiter==0.9.0
pyfcm==1.0.7
distro==1.9.0

# === WebSockets / Async ===
websockets==10.4

# === Natural Language Processing ===
transformers==4.30.2
torch==2.6.0
sentencepiece>=0.1.97
accelerate>=0.18.0
nltk==3.8.1
